Refactors upload method selection to shared component
