+++ client/src/components/Panels/Upload/UploadPage.vue
@@ -1,0 +1,21 @@
+<script setup lang="ts">
+import UploadMethodList from "./UploadMethodList.vue";
+import BreadcrumbHeading from "@/components/Common/BreadcrumbHeading.vue";
+
+const breadcrumbItems = [{ title: "Import Data" }];
+</script>
+
+<template>
+    <div class="upload-page h-100 d-flex flex-column">
+        <BreadcrumbHeading :items="breadcrumbItems" />
+
+        <div class="upload-page-content flex-grow-1 overflow-auto px-3 pt-3">
+            <div class="mb-3">
+                <h2 class="h-lg mb-2">Choose an Import Method</h2>
+                <p class="text-muted">Select how you want to import data into Galaxy</p>
+            </div>
+
+            <UploadMethodList :in-panel="false" />
+        </div>
+    </div>
+</template>
